requirements.txt
alembic==1.0.0
aniso8601==3.0.2
APScheduler==3.5.1
asn1crypto==0.24.0
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
click==6.7
cryptography==2.3
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSE==0.2.1
gevent==1.3.5
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.5
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.10
tzlocal==1.5.1
Werkzeug==0.14.1
